% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{db_con_p}
\alias{db_con_p}
\title{Connect to the intendo database using stored credentials plus a password}
\usage{
db_con_p(password = askpass::askpass())
}
\arguments{
\item{password}{By default, you will be asked for the password.}
}
\description{
The \code{db_con_p()} function is much like the \code{\link[=db_con]{db_con()}} function except that it
requires a password to be supplied at every use. The purpose of the function
is to create a connection object to the \strong{intendo} database.
}
