% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{get_arpu}
\alias{get_arpu}
\title{Get the ARPU value over a particular set of dates}
\usage{
get_arpu(con, dates)
}
\arguments{
\item{con}{A connection to the \strong{intendo} database. We can use either of the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions to generate this.}

\item{dates}{A vector of dates.}
}
\description{
This function uses data from the \strong{intendo} database to get the ARPU
(average revenue per user) on a single day or over a set of days. You need to
supply the database connection object and a vector of dates in ISO format.
}
\examples{
# Create a DB connection
intendo <- db_con()

# Get the ARPU over a three-day period
get_arpu(
  con = intendo,
  dates = c("2015-01-01", "2015-01-02", "2015-01-03")
)

}
