% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_functions}
\alias{tbl_functions}
\alias{tbl_daily_users}
\alias{tbl_revenue}
\alias{tbl_users}
\title{Create \code{tbl} objects for the intendo DB tables}
\usage{
tbl_daily_users(con)

tbl_revenue(con)

tbl_users(con)
}
\arguments{
\item{con}{the name of the connection object.}
}
\description{
The \code{tbl_daily_users()}, \code{tbl_revenue()}, and \code{tbl_users()} functions will
give use \code{tbl_dbi} objects that can be used with \strong{dplyr}. We can use the
\strong{dplyr} API to work with our data and opt to \code{dplyr::collect()} any
summary tables we produce from them. The only requirement is to provide these
functions with a connection object, and this will be \code{intendo} (if we use the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions from this package).
}
\examples{
# Create a DB connection
intendo <- db_con()

# Get a `tbl_dbi` object for the
# `users` DB table
tbl_users(con = intendo)

}
